// src/app/models/PokemonEvolutionChain.ts
// Generated by https://quicktype.io

import { NamedAPIResource } from './NamedAPIResource';

export interface PokemonEvolutionChain {
  baby_trigger_item: NamedAPIResource | null;
  chain: ChainLink;
  id: number;
}

export interface ChainLink {
  evolution_details: EvolutionDetail[];
  evolves_to: ChainLink[];
  is_baby: boolean;
  species: NamedAPIResource;
}

export interface EvolutionDetail {
  gender: number | null;
  held_item: NamedAPIResource | null;
  item: NamedAPIResource | null;
  known_move: NamedAPIResource | null;
  known_move_type: NamedAPIResource | null;
  location: NamedAPIResource | null;
  min_affection: number | null;
  min_beauty: number | null;
  min_happiness: number | null;
  min_level: number | null;
  needs_overworld_rain: boolean;
  party_species: NamedAPIResource | null;
  party_type: NamedAPIResource | null;
  relative_physical_stats: number | null;
  time_of_day: string;
  trade_species: NamedAPIResource | null;
  trigger: NamedAPIResource | null;
  turn_upside_down: boolean;
}